split up helper functions into seperate files
