+++ database/helpers/organizationsModel.js
@@ -1,0 +1,13 @@
+const db = require('../dbConfig')
+
+const getOrgs = async () => {}
+const addOrg = async () => {}
+const updateOrg = async orgId => {}
+const deleteOrg = async orgId => {}
+
+module.exports = {
+  getOrgs,
+  addOrg,
+  updateOrg,
+  deleteOrg
+}
